requirements.txt
Babel==0.9.6
Django==1.4.10
MySQL-python==1.2.4
PIL==1.1.7
Pillow==1.7.8
South==0.7.6
Unidecode==0.04.14
distribute==0.6.27
django-admin-tools==0.5.1
django-autoslug==1.7.1
django-bootstrap3==3.3.0
django-ckeditor==4.0.2
django-compressor==1.3
django-dajax==0.9.2
django-dajaxice==0.6
django-flatblocks==0.8
django-haystack==2.0.0-beta
django-mptt==0.6.0
python-memcached==1.48
sorl-thumbnail==11.12
wsgiref==0.1.2
